+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
+## [26.3.5] - 2026-04-04
+
+### Added
+
+- Added props to button
+
+### Changed
+
+- `themeClass` to method to get class name for another theme
+- Dynamic theme class editing when changing component theme, instead of rerendering
+
+
+
 ## [26.3.4] - 2026-04-04
 
 ### Added